Removes @Transactional on test methods
Removes @Transactional on some test methods to avoid hiding bugs by
reusing in-memory entity objects instead of reloading entities from
the database. This is especially problematic for the following sequence:
creating an entity, then changing it by an api call and finally checking
the result of the change with some asserts.